Combine pool features and separate
* Merged the additions from prototype/pool_siege.rb and associates into
the pooling dir (wasn't much: workload_size, profile_id_generation,
worker_pool_size).
* Removed pool test files from try/prototype so that the prototype and
pooling dir names are accurate to their purpose.
